V2 shadow sample and usage guide
